Add image for musllinux_1_1 policy
The image is based on Alpine Linux 3.12.

+++ docker/build_scripts/install-libxcrypt.sh
@@ -19,21 +19,24 @@ check_sha256sum "v${LIBXCRYPT_VERSION}.tar.gz" "${LIBXCRYPT_HASH}"
 tar xfz "v${LIBXCRYPT_VERSION}.tar.gz"
 pushd "libxcrypt-${LIBXCRYPT_VERSION}"
 ./autogen.sh > /dev/null
+
 DESTDIR=/manylinux-rootfs do_standard_install \
-	--disable-obsolete-api \
-	--enable-hashes=all \
-	--disable-werror
-# we also need libcrypt.so.1 with glibc compatibility for system libraries
-# c.f https://github.com/pypa/manylinux/issues/305#issuecomment-625902928
-make clean > /dev/null
-sed -r -i 's/XCRYPT_([0-9.])+/-/g;s/(%chain OW_CRYPT_1.0).*/\1/g' lib/libcrypt.map.in
-DESTDIR=/manylinux-rootfs/so.1 do_standard_install \
-	--disable-xcrypt-compat-files \
-	--enable-obsolete-api=glibc \
-	--enable-hashes=all \
-	--disable-werror
-cp -P /manylinux-rootfs/so.1/usr/local/lib/libcrypt.so.1* /manylinux-rootfs/usr/local/lib/
-rm -rf /manylinux-rootfs/so.1
+    --disable-obsolete-api \
+    --enable-hashes=all \
+    --disable-werror
+if [ "${POLICY}" != "musllinux_1_1" ]; then 
+    # we also need libcrypt.so.1 with glibc compatibility for system libraries
+    # c.f https://github.com/pypa/manylinux/issues/305#issuecomment-625902928
+    make clean > /dev/null
+    sed -r -i 's/XCRYPT_([0-9.])+/-/g;s/(%chain OW_CRYPT_1.0).*/\1/g' lib/libcrypt.map.in
+    DESTDIR=/manylinux-rootfs/so.1 do_standard_install \
+        --disable-xcrypt-compat-files \
+        --enable-obsolete-api=glibc \
+        --enable-hashes=all \
+        --disable-werror
+    cp -P /manylinux-rootfs/so.1/usr/local/lib/libcrypt.so.1* /manylinux-rootfs/usr/local/lib/
+    rm -rf /manylinux-rootfs/so.1
+fi
 popd
 rm -rf "v${LIBXCRYPT_VERSION}.tar.gz" "libxcrypt-${LIBXCRYPT_VERSION}"
 
@@ -49,4 +52,4 @@ rm -rf /manylinux-rootfs
 # Delete GLIBC version headers and libraries
 rm -rf /usr/include/crypt.h
 find /lib* /usr/lib* \( -name 'libcrypt.a' -o -name 'libcrypt.so' -o -name 'libcrypt.so.*' -o -name 'libcrypt-2.*.so' \) -delete
-ldconfig
+ldconfig /
